@@ -6,7 +6,7 @@ authors:
 homepage: https://github.com/edufolly/flutter_bluetooth_serial/tree/master/example/
 
 environment:
-  sdk: ">=2.12.0 <3.0.0"
+  sdk: ">=2.17.0 <4.0.0"
 
 dependencies:
   flutter:
