@@ -6,12 +6,13 @@ repository: https://github.com/edufolly/flutter_bluetooth_serial
 issue_tracker: https://github.com/edufolly/flutter_bluetooth_serial/issues
 
 environment:
-  sdk: '>=2.12.0 <3.0.0'
-  flutter: ">=1.17.0"
+  sdk: '>=2.17.0 <4.0.0'
+  flutter: ">=2.10.0"
 
 dependencies:
   flutter:
     sdk: flutter
+  permission_handler: ^11.3.1
 
 flutter:
   plugin:
