@@ -3,7 +3,8 @@ library flutter_bluetooth_serial;
 import 'dart:async';
 import 'dart:typed_data';
 import 'dart:convert';
-
+import 'dart:io';
+import 'package:permission_handler/permission_handler.dart';
 import 'package:flutter/services.dart';
 
 part './BluetoothState.dart';
